# run compilation of tests in parallel