--- a/projects/com.oracle.truffle.llvm.parser.base/src/com/oracle/truffle/llvm/parser/base/model/blocks/InstructionBlock.java
+++ b/projects/com.oracle.truffle.llvm.parser.base/src/com/oracle/truffle/llvm/parser/base/model/blocks/InstructionBlock.java
@@ -257,21 +257,7 @@ public final class InstructionBlock implements InstructionGenerator, ValueSymbol
 
     @Override
     public void setName(String name) {
-        // we align the naming to the textparser to make lifetime analysis tests work
-        if (isNumber(name)) {
-            this.name = String.format("%%\"%s\"", name);
-        } else {
-            this.name = String.format("%%%s", name);
-        }
-    }
-
-    private static boolean isNumber(String str) {
-        for (char c : str.toCharArray()) {
-            if (!Character.isDigit(c)) {
-                return false;
-            }
-        }
-        return true;
+        this.name = String.format("%%%s", name);
     }
 
     @Override