--- a/projects/com.oracle.truffle.llvm.runtime/src/com/oracle/truffle/llvm/runtime/LLVMLogger.java
+++ b/projects/com.oracle.truffle.llvm.runtime/src/com/oracle/truffle/llvm/runtime/LLVMLogger.java
@@ -69,4 +69,12 @@ public final class LLVMLogger {
         }
     }
 
+    public static void debug(String string) {
+        CompilerAsserts.neverPartOfCompilation();
+        if (LLVMOptions.DEBUG.verbose()) {
+            // Checkstyle: stop
+            System.err.println(string);
+            // Checkstyle: resume
+        }
+    }
 }