--- a/projects/com.oracle.truffle.llvm.test/src/com/oracle/truffle/llvm/test/LLVMMainArgTestSuite.java
+++ b/projects/com.oracle.truffle.llvm.test/src/com/oracle/truffle/llvm/test/LLVMMainArgTestSuite.java
@@ -94,7 +94,7 @@ public class LLVMMainArgTestSuite extends AbstractMainArgsTestBase<MainArgsTests
     }
 
     @Override
-    protected TestCaseFiles getTestCaseFiles(MainArgsTests prog) {
+    protected TestCaseFile getTestCaseFiles(MainArgsTests prog) {
         return TestHelper.compileToLLVMIRWithClang(prog.getFile(),
                         TestHelper.getTempLLFile(prog.getFile(), "_main"), prog.getFlags(),
                         ClangOptions.builder().optimizationLevel(OptimizationLevel.NONE));