--- a/projects/com.oracle.truffle.llvm.test/src/com/oracle/truffle/llvm/test/TestGCCCompileSuite.java
+++ b/projects/com.oracle.truffle.llvm.test/src/com/oracle/truffle/llvm/test/TestGCCCompileSuite.java
@@ -46,18 +46,18 @@ import com.oracle.truffle.llvm.runtime.LLVMLogger;
 @RunWith(Parameterized.class)
 public class TestGCCCompileSuite extends TestSuiteBase {
 
-    private TestCaseFiles tuple;
+    private TestCaseFile tuple;
 
-    public TestGCCCompileSuite(TestCaseFiles tuple) {
+    public TestGCCCompileSuite(TestCaseFile tuple) {
         this.tuple = tuple;
     }
 
     @Parameterized.Parameters
-    public static List<TestCaseFiles[]> getTestFiles() throws IOException {
+    public static List<TestCaseFile[]> getTestFiles() throws IOException {
         File configFile = LLVMPaths.GCC_TEST_SUITE_COMPILE_TORTURE_CONFIG;
         File testSuite = LLVMPaths.GCC_TEST_SUITE_COMPILE_TORTURE;
         LLVMLogger.info("...start to read and compile files");
-        List<TestCaseFiles[]> files = getTestCasesFromConfigFile(configFile, testSuite, new TestCaseGeneratorImpl(true));
+        List<TestCaseFile[]> files = getTestCasesFromConfigFile(configFile, testSuite, new TestCaseGeneratorImpl(true));
         LLVMLogger.info("...finished reading and compiling files!");
         return files;
     }