--- a/projects/com.oracle.truffle.llvm.test/src/com/oracle/truffle/llvm/test/TestSuiteBase.java
+++ b/projects/com.oracle.truffle.llvm.test/src/com/oracle/truffle/llvm/test/TestSuiteBase.java
@@ -196,7 +196,7 @@ public abstract class TestSuiteBase {
         }
 
         /**
-         * Compile File with given specifications
+         * Compile File with given specifications.
          */
         @Override
         public List<TestCaseFile> getCompiledTestCaseFiles(SpecificationEntry toBeCompiled) {