Based on: #552

#### My performance measurement results:

CPU: Intel(R) Core(TM) i7-6700HQ CPU @ 2.60GHz (4+4 Cores)

TESTCASE: ```mx su-tests lifetime```

OLD: 51758ms

NEW:
1 Thread: 52053ms
2 Threads: 28494ms
4 Threads: 17088ms
8 Threads: 12473ms
16 Threads: 16598ms
32 Threads: 21349ms