@pointhi 2016-11-23
Ok, after talking with @grimmerm This change is no longer needed.